Attempt to implement active selection in the list.
This is a struggle. The state management will allow the updating of a property in the planet object array.
But it will not trigger a re-render of the name in the list. Single tracked values are not a problem.
Need to hunt down what the limitation is for object arrays in a service.